Match String.format rounding in formatFloat
formatFloat rounded with roundToLong() on a Float product, which rounds
ties toward +infinity and manufactures .5 ties that the decimal value
never had; "%.Nf" on the JVM rounds half away from zero on the value's
decimal expansion. Fuel trim 0x08 (-93.75) came out "-93.7" instead of
"-93.8", and module voltage 0x270B (9.99499988) came out "10.00" instead
of "9.99".

Round in Double with explicit half-away-from-zero and an integer divisor.
Add a Double overload so FuelConsumptionRate (x0.05) and FuelRailPressure
(x0.079) format the Double they compute, as master did, instead of
truncating to Float first (0x10E3 * 0.05 = 216.15 -> "216.1" after the
truncation, "216.2" on master).

Add UtilsTest with literal expectations captured from JVM String.format,
since the command tests derive their expected strings from formatFloat
itself.

+++ src/commonMain/kotlin/com/github/eltonvs/obd/command/fuel/Fuel.kt
@@ -43,7 +43,7 @@ class FuelConsumptionRateCommand : ObdCommand() {
 
     override val defaultUnit = "L/h"
     override val handler = { response: ObdRawResponse ->
-        formatFloat((bytesToInt(response.bufferedValue) * FUEL_CONSUMPTION_FACTOR).toFloat(), 1)
+        formatFloat(bytesToInt(response.bufferedValue) * FUEL_CONSUMPTION_FACTOR, 1)
     }
 }
